Move shared logic for generating customer wishlist into hook
