+++ packages/peregrine/lib/talons/Gallery/gallery.gql.ee.js
@@ -1,8 +1,5 @@
 import { gql } from '@apollo/client';
 
-import { GET_PRODUCTS_IN_WISHLISTS } from '../Wishlist/GalleryButton/galleryButton.gql';
-import { GET_WISHLIST_ITEMS } from './gallery.gql.ce';
-
 export const GET_WISHLIST_CONFIG = gql`
     query GetWishlistConfigForGalleryEE {
         storeConfig {
@@ -14,7 +11,5 @@ export const GET_WISHLIST_CONFIG = gql`
 `;
 
 export default {
-    getProductsInWishlistsQuery: GET_PRODUCTS_IN_WISHLISTS,
-    getWishlistConfigQuery: GET_WISHLIST_CONFIG,
-    getWishlistItemsQuery: GET_WISHLIST_ITEMS
+    getWishlistConfigQuery: GET_WISHLIST_CONFIG
 };
